Implements parser with bufio, test, a input generator and results parser. Closes #2

+++ dp_mst/internal/common/common.go
@@ -16,7 +16,7 @@ const (
 )
 
 type Edge struct {
-	X, Y int32
+	X, Y int64
 	W    float64
 }
